Creaton of modified Keymap. Add rule for when holding the a button longer, ä should appear. Add Unicode support for Windows

+++ quantum/unicode/unicode.c
@@ -66,7 +66,7 @@ uint8_t          unicode_saved_mods;
 led_t            unicode_saved_led_state;
 
 #if UNICODE_SELECTED_MODES != -1
-static uint8_t selected[]     = {UNICODE_SELECTED_MODES};
+static uint16_t selected[]     = {UNICODE_SELECTED_MODES};
 static int8_t  selected_count = ARRAY_SIZE(selected);
 static int8_t  selected_index;
 #endif
